refactor: move examples to separate package for better ts support
