+++ docs/package.json
@@ -12,7 +12,6 @@
     "prebuild": "dxtype src/stories/proto/TaskList.proto src/stories/proto/gen/TaskList.ts",
     "build": "tsc && tsc -p src/stories/react/tsconfig.json && tsc -p src/templates/vite/tsconfig.json",
     "bundle": "vuepress build docs",
-    "chromatic": "chromatic --only-changed --project-token=\"44ade288043a\" --storybook-build-dir src/stories/react/out --exit-zero-on-changes",
     "clean:api": "rm -rf docs/api",
     "prerelease": "pnpm typedoc && pnpm clean:api && pnpm apidoc",
     "serve": "vuepress dev docs",
@@ -28,7 +27,7 @@
     "@dxos/client-services": "workspace:*",
     "@dxos/config": "workspace:*",
     "@dxos/echo-schema": "workspace:*",
-    "@dxos/echo-typegen": "workspace:*",
+    "@dxos/examples": "workspace:*",
     "@dxos/keys": "workspace:*",
     "@dxos/log": "workspace:*",
     "@dxos/protocols": "workspace:*",
@@ -51,7 +50,6 @@
     "@vuepress/plugin-register-components": "2.0.0-beta.61",
     "@vuepress/plugin-search": "2.0.0-beta.61",
     "change-case": "^4.1.2",
-    "chromatic": "^6.17.0",
     "codesandbox-import-utils": "^2.2.3",
     "eslint": "^8.41.0",
     "front-matter": "^4.0.2",
