--- docs/src/stories/react/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "lib": [
-      "DOM",
-      "ESNext"
-    ],
-    "module": "ESNext",
-    "rootDir": "..",
-    "outDir": "dist",
-    "types": [
-      "node"
-    ]
-  },
-  "include": [
-    "proto/**/*.ts",
-    "react/**/*.ts"
-  ],
-  "exclude": [
-    "dist",
-    "node_modules",
-    "out"
-  ],
-  "references": []
-}
