+++ packages/sdk/examples/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@dxos/examples",
+  "version": "0.1.45",
+  "private": true,
+  "description": "Example stories used by the docs.",
+  "homepage": "https://dxos.org",
+  "bugs": "https://github.com/dxos/dxos/issues",
+  "license": "MIT",
+  "author": "info@dxos.org",
+  "type": "module",
+  "scripts": {
+    "chromatic": "chromatic --only-changed --project-token=\"44ade288043a\" --storybook-build-dir out --exit-zero-on-changes",
+    "prebuild": "dxtype src/proto/TaskList.proto src/proto/gen/TaskList.ts"
+  },
+  "devDependencies": {
+    "@dxos/aurora": "workspace:*",
+    "@dxos/aurora-theme": "workspace:*",
+    "@dxos/echo-schema": "workspace:*",
+    "@dxos/echo-typegen": "workspace:*",
+    "@dxos/react-appkit": "workspace:*",
+    "@dxos/react-client": "workspace:*",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^3.0.1",
+    "chromatic": "^6.17.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tailwindcss": "~3.3.2",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.9"
+  }
+}
