+++ packages/sdk/examples/src/template/dx.yml
