+++ packages/sdk/examples/src/template/src/components/Demo.tsx
