+++ packages/sdk/examples/src/template/src/components/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export { default as Demo } from './Demo';
