+++ packages/sdk/examples/src/template/src/main.tsx
@@ -2,7 +2,7 @@
 // Copyright 2022 DXOS.org
 //
 
-import React, { StrictMode } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import { setupPeersInSpace } from '@dxos/react-client/testing';
@@ -13,9 +13,7 @@ const main = async () => {
   const { spaceKey, clients } = await setupPeersInSpace({ count: 2 });
 
   ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-    <StrictMode>
-      <App spaceKey={spaceKey} clients={clients} />
-    </StrictMode>,
+    <App spaceKey={spaceKey} clients={clients} />,
   );
 };
