+++ packages/sdk/examples/src/template/src/proto.ts
