+++ packages/sdk/examples/src/template/src/proto/index.ts
