+++ packages/sdk/examples/src/template/tsconfig.json
