+++ packages/sdk/examples/src/template/tsconfig.node.json
