+++ packages/sdk/examples/src/template/vite.config.ts
